Block Editor: Move useResizeCanvas to folder by name
