+++ packages/block-editor/src/components/use-resize-canvas/index.js
